DCM_group9/requirements.txt
tk
Pillow
pyserial
pytest
numpy
matplotlib